update email validation
